Moved tests to proper place
