+++ src/main/java/com/iota/iri/service/API.java
@@ -628,7 +628,7 @@ public class API {
         }
     }
     
-    public List<Hash> getTransactionToApproveTips(int depth, Optional<Hash> reference) throws Exception{
+    List<Hash> getTransactionToApproveTips(int depth, Optional<Hash> reference) throws Exception{
         if (invalidSubtangleStatus()) {
             throw new IllegalStateException("This operations cannot be executed: The subtangle has not been updated yet.");
         }
